Added score Refactored Places DTO

+++ backend/src/main/java/studies/project/tripadvisor/entity/Place.java
@@ -44,7 +44,7 @@ public class Place {
     private String country;
 
     @Column(name = "ENTRANCE_FEE")
-    private float entranceFee;
+    private Double entranceFee;
 
     @Column(name = "EMAIL")
     private String email;
@@ -60,5 +60,4 @@ public class Place {
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "place")
     @JsonManagedReference
     private Set<Comment> comments;
-
 }
